feat: add more utils and options for component testing as well
Fix types as we no longer care about the return value of test/exec functions.
