+++ cypress/plugins/index.ts
@@ -12,11 +12,10 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
+import { cypressComponentTesting } from '@tablecheck/scripts/cypressComponentTesting';
+
 /**
  * @type {Cypress.PluginConfig}
  */
-// eslint-disable-next-line  @typescript-eslint/no-unused-vars
-module.exports = (on: any, config: any) => {
-  // `on` is used to hook into various events Cypress emits
-  // `config` is the resolved Cypress config
-};
+module.exports = ((on, config) =>
+  cypressComponentTesting(on, config)) as Cypress.PluginConfig;
