+++ package.json
@@ -37,35 +37,44 @@
     "start": "npm run test:watch",
     "start:test-skeleton": "node ./scripts/start.js",
     "test": "start-server-and-test start:test-skeleton http://localhost:8080/ cypress:run",
+    "test:component": "cypress run-ct",
+    "test:component:debug": "cypress run-ct --headed --no-exit",
+    "test:component:watch": "cypress open-ct",
     "test:watch": "start-server-and-test start:test-skeleton http://localhost:8080/ cypress:open",
     "tsc": "tablecheck-scripts tsc",
     "tsc:watch": "tablecheck-scripts tsc -w"
   },
+  "dependencies": {
+    "@types/react": "17.0.37"
+  },
   "peerDependencies": {
-    "@xstate/test": "^0.4.2",
+    "@xstate/test": ">=0.4 <0.6",
     "cypress": "^8 || ^9",
     "xstate": "^4"
   },
   "devDependencies": {
-    "@auto-it/all-contributors": "10.32.0",
-    "@auto-it/conventional-commits": "10.32.0",
-    "@auto-it/first-time-contributor": "10.32.0",
-    "@auto-it/npm": "10.32.0",
-    "@auto-it/released": "10.32.0",
-    "@auto-it/slack": "10.32.0",
-    "@commitlint/cli": "13.1.0",
+    "@auto-it/all-contributors": "10.32.3",
+    "@auto-it/conventional-commits": "10.32.3",
+    "@auto-it/first-time-contributor": "10.32.3",
+    "@auto-it/npm": "10.32.3",
+    "@auto-it/released": "10.32.3",
+    "@auto-it/slack": "10.32.3",
+    "@commitlint/cli": "13.2.1",
+    "@cypress/react": "5.10.3",
+    "@cypress/webpack-dev-server": "1.7.0",
     "@tablecheck/commitlint-config": "1.0.2",
-    "@tablecheck/scripts": "1.5.1",
-    "@xstate/test": "0.4.2",
-    "auditjs": "4.0.32",
-    "auto": "10.32.0",
+    "@tablecheck/scripts": "1.8.2",
+    "@xstate/react": "1.6.3",
+    "@xstate/test": "0.5.0",
+    "auditjs": "4.0.33",
+    "auto": "10.32.3",
     "cypress": "9.1.1",
-    "husky": "7.0.2",
+    "husky": "7.0.4",
     "live-server": "1.2.1",
-    "prettier": "2.4.1",
+    "prettier": "2.5.1",
     "start-server-and-test": "1.14.0",
     "todomvc-app-css": "2.4.1",
     "todomvc-common": "1.0.5",
-    "xstate": "4.23.1"
+    "xstate": "4.26.1"
   }
 }
